backend/pyproject.toml
[project]
name = "hotels-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "django-phonenumber-field>=8.4.0",
    "djangorestframework",
    "faker>=40.13.0",
    "phonenumberslite>=9.0.27",
    "pillow>=12.2.0",
    "psycopg2-binary",
    "python-decouple",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "ruff",
]
